+++ flask_unchained/bundles/controller/utils.py
@@ -211,6 +211,14 @@ def redirect(where: Optional[str] = None,
     return flask_redirect('/')
 
 
+def rename_parent_resource_param_name(route, rule: str) -> str:
+    param_to_rename = route._parent_resource_cls._member_param
+    type_, orig_name = get_param_tuples(param_to_rename)[0]
+    orig_param = f'<{type_}:{orig_name}>'
+    renamed_param = f'<{type_}:{controller_name(route._parent_resource_cls)}_{orig_name}>'
+    return rule.replace(orig_param, renamed_param, 1)
+
+
 def _missing_to_default(arg, default=None):
     return arg if arg is not _missing else default