+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 * add a `_form_name` attribute to `FlaskForm` to automatically name forms when rendering them programmatically
 * add `get_extension_local_proxy` and `get_service_local_proxy` methods to the `Unchained` extension
 * add support for overriding static files from bundles
+* minor refactor of the declarative routing for `Controller` and `Resource` classes
+   - consolidate default route rule generation into the `Route` class
+   - make it possible to override the `member_param` of a `Resource` with the `resource` routes function
 
 ## 0.5.1 (2018/07/25)